requirements/dev.txt
# requirements/dev.txt
-r common.txt
django-debug-toolbar==1.2.2
lettuce==0.2.20
